Restructure alerts routes: incidents-first with rules sub-route
/alerts now lands on a new Incidents view (status filter with an
"Open" default covering acknowledged + unacknowledged firing events,
compact metric cards, 15s polling) while rules move to /alerts/rules
and the edit builder to /alerts/rules/$ruleId. Legacy ?tab=rules and
/alerts/$ruleId deep links redirect. The 549-line workspace component
is split into connection-incidents-view, connection-rules-view, and a
segmented alerts-view-switcher; the hero gradient section is gone.
/alerts/new gains a pass-through template/from search schema for the
upcoming builder work.

Also repairs two pre-existing red checks on this branch: the summary
fixture types in use-alerts.test.tsx (new firing/acknowledged/open
shape) and missing webhook-destination hook mocks in settings.test.tsx.
