+++ apps/web/src/routeTree.gen.ts
@@ -45,6 +45,8 @@ import { Route as OrgSlugCConnectionIdAlertsIndexRouteImport } from './routes/$o
 import { Route as OrgSlugCConnectionIdQueuesQueueNameRouteImport } from './routes/$orgSlug.c.$connectionId.queues.$queueName'
 import { Route as OrgSlugCConnectionIdAlertsNewRouteImport } from './routes/$orgSlug.c.$connectionId.alerts.new'
 import { Route as OrgSlugCConnectionIdAlertsRuleIdRouteImport } from './routes/$orgSlug.c.$connectionId.alerts.$ruleId'
+import { Route as OrgSlugCConnectionIdAlertsRulesIndexRouteImport } from './routes/$orgSlug.c.$connectionId.alerts.rules.index'
+import { Route as OrgSlugCConnectionIdAlertsRulesRuleIdRouteImport } from './routes/$orgSlug.c.$connectionId.alerts.rules.$ruleId'
 import { Route as OrgSlugCConnectionIdQueuesQueueNameScheduledJobsNewRouteImport } from './routes/$orgSlug.c.$connectionId.queues.$queueName_.scheduled-jobs.new'
 import { Route as OrgSlugCConnectionIdQueuesQueueNameScheduledJobsSchedulerIdRouteImport } from './routes/$orgSlug.c.$connectionId.queues.$queueName_.scheduled-jobs.$schedulerId'
 import { Route as OrgSlugCConnectionIdQueuesQueueNameJobsJobIdRouteImport } from './routes/$orgSlug.c.$connectionId.queues.$queueName_.jobs.$jobId'
@@ -243,6 +245,18 @@ const OrgSlugCConnectionIdAlertsRuleIdRoute =
     path: '/$ruleId',
     getParentRoute: () => OrgSlugCConnectionIdAlertsRoute,
   } as any)
+const OrgSlugCConnectionIdAlertsRulesIndexRoute =
+  OrgSlugCConnectionIdAlertsRulesIndexRouteImport.update({
+    id: '/rules/',
+    path: '/rules/',
+    getParentRoute: () => OrgSlugCConnectionIdAlertsRoute,
+  } as any)
+const OrgSlugCConnectionIdAlertsRulesRuleIdRoute =
+  OrgSlugCConnectionIdAlertsRulesRuleIdRouteImport.update({
+    id: '/rules/$ruleId',
+    path: '/rules/$ruleId',
+    getParentRoute: () => OrgSlugCConnectionIdAlertsRoute,
+  } as any)
 const OrgSlugCConnectionIdQueuesQueueNameScheduledJobsNewRoute =
   OrgSlugCConnectionIdQueuesQueueNameScheduledJobsNewRouteImport.update({
     id: '/queues/$queueName_/scheduled-jobs/new',
@@ -301,6 +315,8 @@ export interface FileRoutesByFullPath {
   '/$orgSlug/c/$connectionId/alerts/new': typeof OrgSlugCConnectionIdAlertsNewRoute
   '/$orgSlug/c/$connectionId/queues/$queueName': typeof OrgSlugCConnectionIdQueuesQueueNameRoute
   '/$orgSlug/c/$connectionId/alerts/': typeof OrgSlugCConnectionIdAlertsIndexRoute
+  '/$orgSlug/c/$connectionId/alerts/rules/$ruleId': typeof OrgSlugCConnectionIdAlertsRulesRuleIdRoute
+  '/$orgSlug/c/$connectionId/alerts/rules/': typeof OrgSlugCConnectionIdAlertsRulesIndexRoute
   '/$orgSlug/c/$connectionId/queues/$queueName/jobs/$jobId': typeof OrgSlugCConnectionIdQueuesQueueNameJobsJobIdRoute
   '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/$schedulerId': typeof OrgSlugCConnectionIdQueuesQueueNameScheduledJobsSchedulerIdRoute
   '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/new': typeof OrgSlugCConnectionIdQueuesQueueNameScheduledJobsNewRoute
@@ -338,6 +354,8 @@ export interface FileRoutesByTo {
   '/$orgSlug/c/$connectionId/alerts/new': typeof OrgSlugCConnectionIdAlertsNewRoute
   '/$orgSlug/c/$connectionId/queues/$queueName': typeof OrgSlugCConnectionIdQueuesQueueNameRoute
   '/$orgSlug/c/$connectionId/alerts': typeof OrgSlugCConnectionIdAlertsIndexRoute
+  '/$orgSlug/c/$connectionId/alerts/rules/$ruleId': typeof OrgSlugCConnectionIdAlertsRulesRuleIdRoute
+  '/$orgSlug/c/$connectionId/alerts/rules': typeof OrgSlugCConnectionIdAlertsRulesIndexRoute
   '/$orgSlug/c/$connectionId/queues/$queueName/jobs/$jobId': typeof OrgSlugCConnectionIdQueuesQueueNameJobsJobIdRoute
   '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/$schedulerId': typeof OrgSlugCConnectionIdQueuesQueueNameScheduledJobsSchedulerIdRoute
   '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/new': typeof OrgSlugCConnectionIdQueuesQueueNameScheduledJobsNewRoute
@@ -380,6 +398,8 @@ export interface FileRoutesById {
   '/$orgSlug/c/$connectionId/alerts/new': typeof OrgSlugCConnectionIdAlertsNewRoute
   '/$orgSlug/c/$connectionId/queues/$queueName': typeof OrgSlugCConnectionIdQueuesQueueNameRoute
   '/$orgSlug/c/$connectionId/alerts/': typeof OrgSlugCConnectionIdAlertsIndexRoute
+  '/$orgSlug/c/$connectionId/alerts/rules/$ruleId': typeof OrgSlugCConnectionIdAlertsRulesRuleIdRoute
+  '/$orgSlug/c/$connectionId/alerts/rules/': typeof OrgSlugCConnectionIdAlertsRulesIndexRoute
   '/$orgSlug/c/$connectionId/queues/$queueName_/jobs/$jobId': typeof OrgSlugCConnectionIdQueuesQueueNameJobsJobIdRoute
   '/$orgSlug/c/$connectionId/queues/$queueName_/scheduled-jobs/$schedulerId': typeof OrgSlugCConnectionIdQueuesQueueNameScheduledJobsSchedulerIdRoute
   '/$orgSlug/c/$connectionId/queues/$queueName_/scheduled-jobs/new': typeof OrgSlugCConnectionIdQueuesQueueNameScheduledJobsNewRoute
@@ -423,6 +443,8 @@ export interface FileRouteTypes {
     | '/$orgSlug/c/$connectionId/alerts/new'
     | '/$orgSlug/c/$connectionId/queues/$queueName'
     | '/$orgSlug/c/$connectionId/alerts/'
+    | '/$orgSlug/c/$connectionId/alerts/rules/$ruleId'
+    | '/$orgSlug/c/$connectionId/alerts/rules/'
     | '/$orgSlug/c/$connectionId/queues/$queueName/jobs/$jobId'
     | '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/$schedulerId'
     | '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/new'
@@ -460,6 +482,8 @@ export interface FileRouteTypes {
     | '/$orgSlug/c/$connectionId/alerts/new'
     | '/$orgSlug/c/$connectionId/queues/$queueName'
     | '/$orgSlug/c/$connectionId/alerts'
+    | '/$orgSlug/c/$connectionId/alerts/rules/$ruleId'
+    | '/$orgSlug/c/$connectionId/alerts/rules'
     | '/$orgSlug/c/$connectionId/queues/$queueName/jobs/$jobId'
     | '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/$schedulerId'
     | '/$orgSlug/c/$connectionId/queues/$queueName/scheduled-jobs/new'
@@ -501,6 +525,8 @@ export interface FileRouteTypes {
     | '/$orgSlug/c/$connectionId/alerts/new'
     | '/$orgSlug/c/$connectionId/queues/$queueName'
     | '/$orgSlug/c/$connectionId/alerts/'
+    | '/$orgSlug/c/$connectionId/alerts/rules/$ruleId'
+    | '/$orgSlug/c/$connectionId/alerts/rules/'
     | '/$orgSlug/c/$connectionId/queues/$queueName_/jobs/$jobId'
     | '/$orgSlug/c/$connectionId/queues/$queueName_/scheduled-jobs/$schedulerId'
     | '/$orgSlug/c/$connectionId/queues/$queueName_/scheduled-jobs/new'
@@ -772,6 +798,20 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof OrgSlugCConnectionIdAlertsRuleIdRouteImport
       parentRoute: typeof OrgSlugCConnectionIdAlertsRoute
     }
+    '/$orgSlug/c/$connectionId/alerts/rules/': {
+      id: '/$orgSlug/c/$connectionId/alerts/rules/'
+      path: '/rules'
+      fullPath: '/$orgSlug/c/$connectionId/alerts/rules/'
+      preLoaderRoute: typeof OrgSlugCConnectionIdAlertsRulesIndexRouteImport
+      parentRoute: typeof OrgSlugCConnectionIdAlertsRoute
+    }
+    '/$orgSlug/c/$connectionId/alerts/rules/$ruleId': {
+      id: '/$orgSlug/c/$connectionId/alerts/rules/$ruleId'
+      path: '/rules/$ruleId'
+      fullPath: '/$orgSlug/c/$connectionId/alerts/rules/$ruleId'
+      preLoaderRoute: typeof OrgSlugCConnectionIdAlertsRulesRuleIdRouteImport
+      parentRoute: typeof OrgSlugCConnectionIdAlertsRoute
+    }
     '/$orgSlug/c/$connectionId/queues/$queueName_/scheduled-jobs/new': {
       id: '/$orgSlug/c/$connectionId/queues/$queueName_/scheduled-jobs/new'
       path: '/queues/$queueName/scheduled-jobs/new'
@@ -824,6 +864,8 @@ interface OrgSlugCConnectionIdAlertsRouteChildren {
   OrgSlugCConnectionIdAlertsRuleIdRoute: typeof OrgSlugCConnectionIdAlertsRuleIdRoute
   OrgSlugCConnectionIdAlertsNewRoute: typeof OrgSlugCConnectionIdAlertsNewRoute
   OrgSlugCConnectionIdAlertsIndexRoute: typeof OrgSlugCConnectionIdAlertsIndexRoute
+  OrgSlugCConnectionIdAlertsRulesRuleIdRoute: typeof OrgSlugCConnectionIdAlertsRulesRuleIdRoute
+  OrgSlugCConnectionIdAlertsRulesIndexRoute: typeof OrgSlugCConnectionIdAlertsRulesIndexRoute
 }
 
 const OrgSlugCConnectionIdAlertsRouteChildren: OrgSlugCConnectionIdAlertsRouteChildren =
@@ -832,6 +874,10 @@ const OrgSlugCConnectionIdAlertsRouteChildren: OrgSlugCConnectionIdAlertsRouteCh
       OrgSlugCConnectionIdAlertsRuleIdRoute,
     OrgSlugCConnectionIdAlertsNewRoute: OrgSlugCConnectionIdAlertsNewRoute,
     OrgSlugCConnectionIdAlertsIndexRoute: OrgSlugCConnectionIdAlertsIndexRoute,
+    OrgSlugCConnectionIdAlertsRulesRuleIdRoute:
+      OrgSlugCConnectionIdAlertsRulesRuleIdRoute,
+    OrgSlugCConnectionIdAlertsRulesIndexRoute:
+      OrgSlugCConnectionIdAlertsRulesIndexRoute,
   }
 
 const OrgSlugCConnectionIdAlertsRouteWithChildren =
